move calendar marks to shared/types folder
