+++ pages/disruption/@id/+data.ts
@@ -4,10 +4,8 @@ import { JsonSerializable } from "@/shared/types/json-serializable";
 import { StandardDisruptionPeriod } from "@/server/data/disruption/period/standard-disruption-period";
 import { EndsExactly } from "@/server/data/disruption/period/ends/ends-exactly";
 import { EveningsOnlyDisruptionPeriod } from "@/server/data/disruption/period/evenings-only-disruption-period";
-import {
-  RenderedCalendarMark,
-  renderCalendarMarks,
-} from "@/server/data/disruption/period/utils/render-calendar-marks";
+import { renderCalendarMarks } from "@/server/data/disruption/period/utils/render-calendar-marks";
+import { RenderedCalendarMark } from "@/shared/types/calendar-marks";
 
 export type Data = {
   backHref: string;
