+++ shared/types/calendar-marks.ts
@@ -1,0 +1,8 @@
+export type CalendarMark = "no-disruption" | "evening-only" | "all-day";
+
+export type RenderedCalendarMark = {
+  year: number;
+  month: number;
+  day: number;
+  mark: CalendarMark;
+};
